+++ lib/td_dd/data_structures/data_structure_relation.ex
@@ -19,6 +19,11 @@ defmodule TdDd.DataStructures.DataStructureRelation do
       :child_id
     ])
     |> validate_required([:parent_id, :child_id])
+    |> check_constraint(
+      :parent_id,
+      name: :avoid_structure_relation_itself,
+      message: "Structure must not have relations with itself"
+    )
   end
 
   @doc false